forecast-engine/requirements.txt
# Statistical Forecasting Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Time Series Models
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Prophet requires specific installation
# Install separately: pip install prophet
# Or use conda: conda install -c conda-forge prophet

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Visualization (optional for testing)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
